reads data from firebase Agenda Screen and update appointments according to new collection

+++ lib/screens/appointment_confirm_detail_screen/appointment_confirm_detail_controller.dart
@@ -89,9 +89,9 @@ class AppointmentConfirmDetailController extends GetxController {
   }
 
   String getTotalPrice(List<Treatment> services) {
-    int price = 0;
+    double price = 0;
     services.forEach((service) {
-      price += int.parse(service.price!);
+      price += double.parse(service.price!);
     });
     return price.toString();
   }
